Working commit - users and groups created
This commit contains the cached data from which to do te lookups
to create the groups and users for the vo pools.

The tasks added are inner loops, which use the vo name as index
and a `set_fact` to create the json_query to use.

There is no proper management of "special" roles yet (prd, sgm)
but the groups are in place to do that. Some feedback is necessary
from the VOs to understand how best to do this.